refactor second pass
