--- src/py3r/behaviour/animation/_projection.py
@@ -1,166 +1,0 @@
-from __future__ import annotations
-
-import numpy as np
-
-
-def _is_valid(pix: np.ndarray, idx: int) -> bool:
-    return idx < len(pix) and pix[idx, 0] >= 0 and pix[idx, 1] >= 0
-
-
-def _compute_bounds(
-    points_xy: np.ndarray,
-    boundary_arrays: list[tuple[str, np.ndarray]],
-    pad: float = 0.05,
-) -> tuple[float, float, float, float]:
-    flat = points_xy.reshape(-1, 2)
-    valid = np.isfinite(flat[:, 0]) & np.isfinite(flat[:, 1])
-    xs = [flat[valid, 0]] if np.any(valid) else []
-    ys = [flat[valid, 1]] if np.any(valid) else []
-    for _, arr in boundary_arrays:
-        poly = np.asarray(arr, dtype=float)
-        if poly.ndim != 3 or poly.shape[2] != 2 or poly.shape[0] == 0 or poly.shape[1] == 0:
-            continue
-        flat_poly = poly.reshape(-1, 2)
-        ok = np.isfinite(flat_poly[:, 0]) & np.isfinite(flat_poly[:, 1])
-        if np.any(ok):
-            xs.append(flat_poly[ok, 0])
-            ys.append(flat_poly[ok, 1])
-    if not xs:
-        return 0.0, 1.0, 0.0, 1.0
-    xmin, xmax = float(np.min(np.concatenate(xs))), float(np.max(np.concatenate(xs)))
-    ymin, ymax = float(np.min(np.concatenate(ys))), float(np.max(np.concatenate(ys)))
-    if xmin == xmax:
-        xmax = xmin + 1.0
-    if ymin == ymax:
-        ymax = ymin + 1.0
-    pad = float(np.clip(pad, 0.0, 0.45))
-    if pad > 0:
-        xspan = xmax - xmin
-        yspan = ymax - ymin
-        xmin -= xspan * pad
-        xmax += xspan * pad
-        ymin -= yspan * pad
-        ymax += yspan * pad
-    return xmin, xmax, ymin, ymax
-
-
-def _coords_to_pixels(
-    coords: np.ndarray,
-    width: int,
-    height: int,
-    bounds: tuple[float, float, float, float],
-    pixel_coords: bool,
-) -> np.ndarray | None:
-    if coords.size == 0:
-        return None
-    arr = np.asarray(coords, dtype=float)
-    out = np.full((arr.shape[0], 2), -1, dtype=np.int32)
-    valid = np.isfinite(arr[:, 0]) & np.isfinite(arr[:, 1])
-    if not np.any(valid):
-        return out
-    x = arr[valid, 0]
-    y = arr[valid, 1]
-    if pixel_coords:
-        xi = np.rint(x).astype(np.int32)
-        yi = np.rint(y).astype(np.int32)
-    else:
-        xmin, xmax, ymin, ymax = bounds
-        sx = max(width - 1, 1) / (xmax - xmin)
-        sy = max(height - 1, 1) / (ymax - ymin)
-        xi = np.rint((x - xmin) * sx).astype(np.int32)
-        yi = np.rint((height - 1) - ((y - ymin) * sy)).astype(np.int32)
-    inb = (xi >= 0) & (xi < width) & (yi >= 0) & (yi < height)
-    idx = np.where(valid)[0][inb]
-    out[idx, 0] = xi[inb]
-    out[idx, 1] = yi[inb]
-    return out
-
-
-def _make_projector(points_xyz: np.ndarray, view: dict | None) -> dict:
-    v = view or {}
-    azim_deg = float(v.get("azim", 45.0))
-    elev_deg = float(v.get("elev", 30.0))
-    proj = str(v.get("proj", "ortho")).lower()
-    if proj not in {"ortho", "persp"}:
-        raise ValueError("view['proj'] must be 'ortho' or 'persp'")
-
-    valid = np.isfinite(points_xyz).all(axis=2)
-    if np.any(valid):
-        center = np.nanmean(points_xyz[valid], axis=0)
-    else:
-        center = np.array([0.0, 0.0, 0.0], dtype=float)
-
-    centered = points_xyz - center
-    az = np.deg2rad(azim_deg)
-    el = np.deg2rad(elev_deg)
-    cz, sz = np.cos(az), np.sin(az)
-    cx, sx = np.cos(el), np.sin(el)
-    rz = np.array([[cz, -sz, 0.0], [sz, cz, 0.0], [0.0, 0.0, 1.0]], dtype=float)
-    rx = np.array([[1.0, 0.0, 0.0], [0.0, cx, -sx], [0.0, sx, cx]], dtype=float)
-    if proj == "ortho":
-        return {"center": center, "rotm": rz.T @ rx.T, "proj": proj}
-
-    finite_vals = centered[np.isfinite(centered)]
-    max_abs = float(np.nanmax(np.abs(finite_vals))) if finite_vals.size > 0 else 1.0
-    camera_distance = float(v.get("camera_distance", max(1.0, 4.0 * max_abs)))
-    focal_length = float(v.get("focal_length", camera_distance))
-    return {
-        "center": center,
-        "rotm": rz.T @ rx.T,
-        "proj": proj,
-        "camera_distance": camera_distance,
-        "focal_length": focal_length,
-    }
-
-
-def _project_xyz_with_projector(points_xyz: np.ndarray, projector: dict) -> np.ndarray:
-    valid = np.isfinite(points_xyz).all(axis=2)
-    centered = points_xyz - projector["center"]
-    rot = centered @ projector["rotm"]
-    x = rot[:, :, 0]
-    y = rot[:, :, 1]
-    z = rot[:, :, 2]
-    if projector["proj"] == "ortho":
-        xy = np.stack((x, y), axis=2)
-        xy[~valid] = np.nan
-        return xy
-    denom = projector["camera_distance"] - z
-    good = valid & np.isfinite(denom) & (denom > 1e-9)
-    xp = np.full_like(x, np.nan, dtype=float)
-    yp = np.full_like(y, np.nan, dtype=float)
-    xp[good] = projector["focal_length"] * x[good] / denom[good]
-    yp[good] = projector["focal_length"] * y[good] / denom[good]
-    return np.stack((xp, yp), axis=2)
-
-
-def _resolve_boundary_z(name: str, boundary_z) -> float:
-    if isinstance(boundary_z, dict):
-        return float(boundary_z.get(name, 0.0))
-    if boundary_z is None:
-        return 0.0
-    return float(boundary_z)
-
-
-def _project_boundary_arrays_3d_to_2d(
-    boundary_arrays: list[tuple[str, np.ndarray]],
-    projector: dict,
-    boundary_z,
-    n_frames: int,
-) -> list[tuple[str, np.ndarray]]:
-    projected: list[tuple[str, np.ndarray]] = []
-    for name, poly in boundary_arrays:
-        arr = np.asarray(poly, dtype=float)
-        if arr.ndim != 3 or arr.shape[0] != n_frames:
-            raise ValueError("Boundary array must have shape (n_frames, n_vertices, 2|3)")
-        if arr.shape[2] == 2:
-            z = _resolve_boundary_z(name, boundary_z)
-            xyz = np.concatenate(
-                (arr, np.full((arr.shape[0], arr.shape[1], 1), z, dtype=float)),
-                axis=2,
-            )
-        elif arr.shape[2] == 3:
-            xyz = arr
-        else:
-            raise ValueError("Boundary array must have 2 or 3 dimensions on last axis")
-        projected.append((str(name), _project_xyz_with_projector(xyz, projector)))
-    return projected
